Moves voice related factories into a folder. Added some comments on voice callbacks. Cleaned up VoiceStateFactory
