+++ DSharpPlus.Voice/Transport/Factories/ITransportServiceBuilder.cs
@@ -1,0 +1,9 @@
+namespace DSharpPlus.Voice.Transport.Factories;
+
+/// <summary>
+/// An abstraction for creating a transport service.
+/// </summary>
+public interface ITransportServiceBuilder
+{
+    public IInitializedTransportServiceBuilder CreateBuilder();
+}
