--- DSharpPlus.Voice/Transport/ITransportServiceBuilder.cs
@@ -1,9 +1,0 @@
-namespace DSharpPlus.Voice.Transport;
-
-/// <summary>
-/// An abstraction for creating a transport service.
-/// </summary>
-public interface ITransportServiceBuilder
-{
-    public IInitializedTransportServiceBuilder CreateBuilder();
-}
