--- DSharpPlus.Voice/Transport/ITransportServiceFactory.cs
@@ -1,12 +1,0 @@
-namespace DSharpPlus.Voice.Transport;
-
-/// <summary>
-/// Provides a DI-friendly mechanism to initialize and create a transport service.
-/// </summary>
-public interface ITransportServiceFactory
-{
-    /// <summary>
-    /// Creates a builder for further initialization of a transport service.
-    /// </summary>
-    public ITransportServiceBuilder CreateTransportService();
-}
